bootcamp_week9/bootcamp_day3/requirements.txt
autogen-agentchat==0.4.0
autogen-ext[openai]==0.4.0
autogen-core==0.4.0
psycopg2-binary
pandas  